refactor: Money의 필드 타입 변경
int -> long

+++ src/main/java/lotto/controller/Controller.java
@@ -44,7 +44,7 @@ public class Controller {
         return new Money(convertStringToInt(money));
     }
 
-    private Lottos getLottos(final int money) {
+    private Lottos getLottos(final long money) {
         return Lottos.createWith(money, new RandomNumbersGenerator());
     }
